+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Changelog
 
+#### Version - TBD
+* Fixed Final Fantasy 7: Remake Intergrade meta data
+
 #### Version - 3.4.1.0 - 12/21/2023
 * Added Support for Final Fantasy 7: Remake Intergrade
 * Update CLI to .NET 8.0 (was missed in the last update)